channel_folders: Move unreads code from left_sidebar_navigation to stream_list.
